move common flow/chooser utility functions to separate file
